+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.11.6 (unreleased)
 
 - Add `sdwan_scope` resource and data source
+- Bump `sdwan_custom_application` schema to `20.18.0`
 
 ## 0.11.5